Refactor `MerakiAPIClient` and fix network parsing bugs
- Moved `get_vlan_data` and `get_network_events` from `MerakiAPIClient` to `NetworkEndpoints` to reduce `client.py` size (God Class refactor).
- Updated `MerakiAPIClient` size is now ~300 lines (was 463).
- Updated `ApplianceFetchStrategy` to use `client.network.get_vlan_data`.
- Fixed bug in `custom_components/meraki_ha/core/parsers/network.py` where `disabled_features` was not updated for specific errors due to incorrect `isinstance` check order.
- Fixed multiple unit tests in `test_data_fetcher.py`, `test_network_parser.py`, and `test_data_fetch_manager_features.py` to align with refactors and fix mock issues.
- Moved client tests to `tests/core/api/endpoints/test_network.py` and deleted `tests/core/api/test_client.py`.
- Verified with `run_checks.sh`.
