+++ custom_components/meraki_ha/core/fetch_strategies/base.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from ..api.client import MerakiAPIClient
