+++ custom_components/meraki_ha/core/parsers/network.py
@@ -5,18 +5,18 @@ from __future__ import annotations
 import logging
 from typing import Any
 
-from ...core.errors import (
-    MerakiInformationalError,
-    MerakiTrafficAnalysisError,
-    MerakiVlanError,
-    MerakiVlansDisabledError,
-)
 from ...core.models.network import (
     MerakiFirewallRule,
     MerakiNetwork,
     MerakiTrafficShaping,
     MerakiVlan,
     MerakiVpn,
+)
+from ..errors import (
+    MerakiInformationalError,
+    MerakiTrafficAnalysisError,
+    MerakiVlanError,
+    MerakiVlansDisabledError,
 )
 
 _LOGGER = logging.getLogger(__name__)
@@ -29,9 +29,7 @@ def parse_network_data(
     disabled_features: set[str],
     coordinator: Any = None,  # Added to support status messages
 ) -> dict[str, Any]:
-    """
-    Parse and process network-level data.
-    """
+    """Parse and process network-level data."""
     appliance_traffic: dict[str, Any] = {}
     vlan_by_network: dict[str, list[MerakiVlan]] = {}
     l3_firewall_rules_by_network: dict[str, list[MerakiFirewallRule]] = {}
@@ -120,8 +118,11 @@ def _parse_traffic(
     """Parse appliance traffic data."""
     key = f"traffic_{network_id}"
     data = detail_data.get(key)
-    
-    if isinstance(data, MerakiInformationalError):
+
+    if isinstance(data, MerakiTrafficAnalysisError):
+        disabled_features.add(key)
+        appliance_traffic[network_id] = {"error": "disabled", "reason": str(data)}
+    elif isinstance(data, MerakiInformationalError):
         if "traffic analysis" in str(data).lower():
             if coordinator:
                 coordinator.add_network_status_message(
@@ -133,9 +134,6 @@ def _parse_traffic(
             "error": "disabled",
             "reason": str(data),
         }
-    elif isinstance(data, MerakiTrafficAnalysisError):
-        disabled_features.add(key)
-        appliance_traffic[network_id] = {"error": "disabled", "reason": str(data)}
     elif isinstance(data, dict):
         appliance_traffic[network_id] = data
     elif previous_data and key in previous_data:
@@ -154,8 +152,12 @@ def _parse_vlans(
     key = f"vlans_{network_id}"
     data = detail_data.get(key)
 
-    if isinstance(data, MerakiInformationalError):
+    if isinstance(data, (MerakiVlanError, MerakiVlansDisabledError)):
+        disabled_features.add(key)
+        vlan_by_network[network_id] = []
+    elif isinstance(data, MerakiInformationalError):
         if "vlans are not enabled" in str(data).lower():
+            disabled_features.add(key)
             if coordinator:
                 coordinator.add_network_status_message(
                     network_id,
@@ -163,9 +165,6 @@ def _parse_vlans(
                 )
                 coordinator.mark_vlan_check_done(network_id)
             vlan_by_network[network_id] = []
-    elif isinstance(data, (MerakiVlanError, MerakiVlansDisabledError)):
-        disabled_features.add(key)
-        vlan_by_network[network_id] = []
     elif isinstance(data, list):
         vlan_by_network[network_id] = [MerakiVlan.from_dict(v) for v in data]
     elif isinstance(data, dict):
